requirements.txt
asgiref==3.10.0
Django==5.2.8
sqlparse==0.5.3
tzdata==2025.2
gunicorn>=21.2.0
django-ninja>=1.5.0
